added signup and signin

+++ app/Http/Controllers/ChirpController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Chirp;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class ChirpController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -36,9 +38,8 @@ class ChirpController extends Controller
             'message.max' => 'Chirps must be 255 characters or less'
         ]);
 
-        Chirp::create([
-            'message' => $validated['message'],
-        ]);
+
+        auth()->user()->chirps()->create($validated);
 
         return redirect('/')->with('success', 'Chirp created!');
     }
@@ -56,7 +57,9 @@ class ChirpController extends Controller
      */
     public function edit(Chirp $chirp)
     {
-        return view('chirps.edit', ['chirp' => $chirp]);
+        $this->authorize('update', $chirp);
+
+        return view('chirps.edit', compact('chirp'));
     }
 
     /**
@@ -64,16 +67,15 @@ class ChirpController extends Controller
      */
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         $validated = $request->validate([
             'message' => 'required|string|max:255',
-        ],[
-            'message.required' =>'Please write something to chirp',
-            'message.max' => 'Chirps must be 255 characters or less'
         ]);
 
         $chirp->update($validated);
 
-        return redirect('/')->with('success', 'Your chirp has been updated!');
+        return redirect('/')->with('success', 'Chirp updated!');
     }
 
     /**
@@ -81,7 +83,10 @@ class ChirpController extends Controller
      */
     public function destroy(Chirp $chirp)
     {
+        $this->authorize('delete', $chirp);
+
         $chirp->delete();
-        return redirect('/')->with('success', 'Your chirp has been deleted!');
+
+        return redirect('/')->with('success', 'Chirp deleted!');
     }
 }
